Logging to file without overhead
There is no need to log to a file using `>> ~/logs/node.out 2>&1 &` and creating additional overhead when there is a standard way to log directly to the file. 

Also, I believe that it is better to log to the usual location `/var/log`, but that is optional I guess.